refactor: Move MetadataService to metadata directory
