+++ frontend/src/app/general/footer/footer.component.ts
@@ -5,7 +5,7 @@
 
 import { Component } from '@angular/core';
 import { MatDialog } from '@angular/material/dialog';
-import { MetadataService } from 'src/app/general/metadata/version/version.service';
+import { MetadataService } from 'src/app/general/metadata/metadata.service';
 
 @Component({
   selector: 'app-footer',
